requirements.txt
pytest
sphinx
sphinx_rtd_theme
black
loguru
hydra-core
pandas
pycaret
openpyxl
pandas_profiling